refactor: coreApi 위치 이동
